seperate camera code from rendering code
