+++ config.py
@@ -1,9 +1,15 @@
+from camera_route import circle_local
+
 # Output figure store place
-pan1 = 'data/'
+pan1 = '.'
+
+# Input file dir
+# map_fname = '/n/fink1/ggreen/bayestar/output/allsky_2MASS/compact/dust-map-3d-uncompressed.h5'#compact_10samp.h5'
+map_fname = 'data/AquilaSouthLarge2_unified.h5' 
 
 # Misc settings
 plot_props = {
-    'fname': pan1 + '3d/allsky_2MASS/AqS/AqS-loop-hq.png', #'3d/allsky_2MASS/grand-tour/simple-loop-att-v2-lq.png',
+    'fname': pan1 + '/3d/allsky_2MASS/AqS/AqS-loop-hq.png', #'3d/allsky_2MASS/grand-tour/simple-loop-att-v2-lq.png',
     'figsize': (10, 7),
     'dpi': 10,
     'n_averaged': 1,
@@ -24,7 +30,7 @@ plot_props = {
 #camera_pos = paper_renderings()
 #camera_pos = Orion_flythrough(n_frames=400)
 #camera_pos = grand_tour_path(n_frames=20)#1600)
-# camera_pos = circle_local(n_frames=40, l_0=30., b_0=5.)
+camera_pos = circle_local(n_frames=1, l_0=30., b_0=5.)
 #camera_pos = stereo_pair()
 
 #for key in camera_pos:
@@ -47,6 +53,4 @@ label_props = {
     'text_color': (0, 166, 255),#(255, 255, 255),
     'stroke_color': (9, 73, 92),#(0, 0, 0) #(192, 225, 235)
 }
-
-# map_fname = '/n/fink1/ggreen/bayestar/output/allsky_2MASS/compact/dust-map-3d-uncompressed.h5'#compact_10samp.h5'
-map_fname = './AquilaSouthLarge2_unified.h5'  
+
